+++ src/Vixen.Modules/App/Curves/CurveEditor.cs
@@ -105,7 +105,7 @@ namespace VixenModules.App.Curves
 			double newX, newY;
 			if (e.Button == MouseButtons.Left && _drawCurve)
 			{
-				int textThreshold = textBoxThreshold.Text == "" ? (int) 0 : Convert.ToInt16(textBoxThreshold.Text);
+				int textThreshold = textBoxThreshold.Text == "" ? 0 : Convert.ToInt16(textBoxThreshold.Text);
 				// only add if we've actually clicked on the pane, so make sure the mouse is over it first
 				if (zedGraphControl.MasterPane.FindPane(e.Location) != null)
 				{
